# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NBE10-12-2-Team1 API
  description: |
    오늘 뭐먹지 SNS 서비스 REST API 문서입니다.

    ## 인증
    인증이 필요한 API는 요청 헤더에 Bearer 토큰을 포함해야 합니다.
    ```
    Authorization: Bearer {accessToken}
    ```
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/reissue:
    post:
      tags: [Auth]
      summary: 토큰 재발급
      description: Refresh Token을 이용해 새로운 Access Token과 Refresh Token을 발급합니다.
      security: []
      responses:
        "200":
          description: 토큰 재발급 성공
          content:
            application/json:
              example:
                success: true
                data: null
        "401":
          description: Refresh Token 만료 / 위조 / 불일치 / 연결된 사용자 없음
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_REFRESH_TOKEN"
                  message: "유효하지 않은 Refresh Token입니다. 다시 로그인해주세요."
        "500":
          description: 서버 오류 / Access Token 재발급 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: 일반 회원가입
      description: 이메일(로그인id)와 비밀번호로 회원가입합니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginId, password, nickname]
              properties:
                loginId:
                  type: string
                  format: email
                  example: "foodie@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "password123!"
                nickname:
                  type: string
                  maxLength: 20
                  example: "푸디보이"
      responses:
        "200":
          description: 회원가입 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  nickname: "푸디보이"
                  profileImage: null
                  email: "foodie@example.com"
                  provider: "LOCAL"
                  createdAt: "2026-07-03T10:00:00"
                message: "회원가입이 완료되었습니다."
        "400":
          description: 필수 필드 누락 / 이메일 형식 오류 / 비밀번호 8자 미만 / 닉네임 20자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "409":
          description: 이미 사용 중인 이메일 (DUPLICATE_EMAIL)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "DUPLICATE_EMAIL"
                  message: "이미 사용 중인 이메일입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 일반 로그인
      description: 이메일(로그인id)와 비밀번호로 로그인합니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginId, password]
              properties:
                loginId:
                  type: string
                  format: email
                  example: "foodie@example.com"
                password:
                  type: string
                  example: "password123!"
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  nickname: "푸디보이"
                  profileImage: null
                  email: "foodie@example.com"
                  provider: "LOCAL"
                  createdAt: "2026-07-03T10:00:00"
                message: "로그인에 성공했습니다."
        "400":
          description: 필수 필드 누락 / 이메일 형식 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "401":
          description: 이메일 또는 비밀번호 불일치 (INVALID_CREDENTIALS)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_CREDENTIALS"
                  message: "이메일 또는 비밀번호가 올바르지 않습니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      description: Access Token을 무효화하여 로그아웃합니다. 카카오 및 일반 로그인 공통으로 사용합니다.
      responses:
        "200":
          description: 로그아웃 성공
          content:
            application/json:
              example:
                success: true
                data: null
                message: "로그아웃되었습니다."
        "401":
          description: 인증되지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "인증이 만료되었습니다. 다시 로그인해주세요."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

  # ─────────────────────────────────────────────
  # Users
  # ─────────────────────────────────────────────
...
  /api/v1/feeds:
    post:
      tags: [Feeds]
      summary: 피드 글 작성
      description: 새로운 피드 글을 작성합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [feed]
              properties:
                feed:
                  type: object
                  required: [content, restaurantId]
                  properties:
                    content:
                      type: string
                      maxLength: 1000
                      example: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                    restaurantId:
                      type: integer
                      example: 1
                image:
                  type: string
                  format: binary
                  description: 피드 이미지 파일 (jpg/jpeg/png, 선택)
      responses:
        "201":
          description: 글 작성 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feedId: 45
                  content: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                  nickname: "푸디보이"
                  profileImage: "https://k.kakaocdn.net/.../img.jpg"
                  likeCount: 0
                  createdAt: "2026-06-18T17:50:00"
                  updatedAt: "2026-06-18T17:50:00"
                  restaurantId: 1
                  restaurantName: "을지삼겹"
                message: "피드가 생성되었습니다."
        "400":
          description: content 없음 또는 1000자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "내용을 입력해주세요. (최대 1000자)"
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "404":
          description: 존재하지 않는 식당 (RESTAURANT_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "RESTAURANT_NOT_FOUND"
                  message: "존재하지 않는 식당입니다."
        "500":
          description: 저장 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    get:
      tags: [Feeds]
      summary: 피드 글 목록 조회
      description: |
        피드 글 목록을 조회합니다.

        **용례:**
        - `GET /api/v1/feeds?restaurantId=1` → 특정 식당 태그된 글들 (식당 페이지)
        - `GET /api/v1/feeds?userId=2` → 특정 유저의 글 (프로필 페이지)
        - `GET /api/v1/feeds` → 전체 공개 글 (탐색)
      parameters:
        - name: restaurantId
          in: query
          required: false
          description: 특정 식당 태그 글만 필터링
          schema:
            type: integer
            example: 1
        - name: userId
          in: query
          required: false
          description: 특정 유저의 글만 필터링
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 글 목록 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feeds:
                    - feedId: 45
                      content: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                      userId: 1
                      nickname: "푸디보이"
                      profileImage: "https://k.kakaocdn.net/.../img.jpg"
                      likeCount: 12
                      commentCount: 4
                      isLikedByMe: true
                      restaurantId: 1
                      restaurantName: "을지삼겹"
                      createdAt: "2026-06-18T17:50:00"
                  totalPages: 3
                  totalElements: 25
                  page: 0
                  size: 10
        "400":
          description: restaurantId/userId 형식 오류 또는 page/size가 0 이하 (INVALID_INPUT_VALUE / INVALID_PAGE_SIZE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
...
  /api/v1/feeds/{id}/comments:
    post:
      tags: [Comments]
      summary: 댓글 작성
      description: 피드 글에 댓글을 작성합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 글 ID
          schema:
            type: integer
            example: 45
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 500
                  example: "저도 여기 자주 가요! 냉면도 맛있어요."
      responses:
        "201":
          description: 댓글 작성 성공
          content:
            application/json:
              example:
                success: true
                data:
                  commentId: 101
                  feedId: 45
                  userId: 2
                  nickname: "서교동김푸디"
                  profileImage: "https://k.kakaocdn.net/.../profile2.jpg"
                  content: "저도 여기 자주 가요! 냉면도 맛있어요."
                  createdAt: "2026-06-18T18:10:00"
                message: "댓글이 작성되었습니다."
        "400":
          description: content 없음 또는 500자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "댓글 내용을 입력해주세요. (최대 500자)"
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "404":
          description: 존재하지 않는 글 (FEED_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FEED_NOT_FOUND"
                  message: "존재하지 않는 게시글입니다."
        "500":
          description: 저장 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    get:
      tags: [Comments]
      summary: 댓글 목록 조회
      description: 피드 글의 댓글 목록을 페이지네이션으로 조회합니다. 댓글이 없으면 빈 배열을 반환합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 글 ID
          schema:
            type: integer
            example: 45
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 20)
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 댓글 목록 조회 성공 (댓글 없을 시 빈 배열 반환)
          content:
            application/json:
              example:
                success: true
                data:
                  comments:
                    - commentId: 101
                      userId: 2
                      nickname: "서교동김푸디"
                      profileImage: "https://k.kakaocdn.net/.../profile2.jpg"
                      content: "저도 여기 자주 가요! 냉면도 맛있어요."
                      createdAt: "2026-06-18T18:10:00"
                  totalPages: 1
                  totalElements: 3
        "400":
          description: page/size가 0 이하 (INVALID_PAGE_SIZE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_PAGE_SIZE"
                  message: "페이지 번호와 크기는 0보다 커야 합니다."
        "404":
          description: 존재하지 않는 글 (FEED_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FEED_NOT_FOUND"
                  message: "존재하지 않는 게시글입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/feeds/{id}/comments/{commentId}:
    delete:
      tags: [Comments]
      summary: 댓글 삭제
      description: 댓글을 삭제합니다. 본인만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 글 ID
          schema:
            type: integer
            example: 45
        - name: commentId
          in: path
          required: true
          description: 댓글 ID
          schema:
            type: integer
            example: 101
      responses:
        "200":
          description: 댓글 삭제 성공
          content:
            application/json:
              example:
                success: true
                data: null
                message: "댓글이 삭제되었습니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 타인 댓글 삭제 시도 (FORBIDDEN_COMMENT_DELETE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN_COMMENT_DELETE"
                  message: "본인의 댓글만 삭제할 수 있습니다."
        "404":
          description: 존재하지 않는 댓글 (COMMENT_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "COMMENT_NOT_FOUND"
                  message: "존재하지 않는 댓글입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

  # ─────────────────────────────────────────────
  # Restaurant Lists
  # ─────────────────────────────────────────────
...